Convert to CMake project
This converts the project to CMake.  One of the major changes is that
libraries (arduinolibs, qzxing, cpptotp, qrcode, scrypt) are now fetched
from their respective github repositories, configured and built rather
than having the code directly copied into this project.  This required
some minor tweaking of some include paths in the source code.

There was also an error in the main.cpp file in which an engine was
declared before an app, which does not work correctly.

Signed-off-by: Ed Beroset <beroset@ieee.org>